backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
mysql-connector-python==8.0.33
python-dotenv==1.0.0
waitress==2.1.2
werkzeug==2.2.2
PyJWT==2.6.0
click==8.1.3